% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_moostr.R
\name{make_moostr}
\alias{make_moostr}
\title{Moodle Syntax String for Multichoice and Numerical Question Types}
\usage{
make_moostr(type, ans, reward, tol = 0.01)
}
\arguments{
\item{type}{string; either "mchoice" or "num"}

\item{ans}{a list of possible answers; for "mchoice" type, each
possible answer is a string; for "num" type, each possible answer is
of numeric type}

\item{reward}{a list of corresponding rewards for each possible
answer expressed in percentages (0 and 100 inclusive)}

\item{tol}{a tolerance level for "num" type; this argument is
ignored for "mchoice" type}
}
\value{
a Moodle syntax string
}
\description{
This package generates a Moodle syntax string for
a given list of possible answers and their corresponding rewards.
It allows the user to conveniently allocate rewards for multiple
correct and partially correct answers For multichoice questions, the
package also randomly shuffles the answer list. For numerical questions,
the package also allows the flexibility to specify different tolerance
levels for each possible answer.

For the package to function properly, it is necessary to set the
question type for both multichoice and numerical questions to vertatim
type.The returned Moodle syntax string can then be directly used as the
solution to a verbatim item in an Embedded Answers (Cloze) exercise
with exams package or Moodle.
}
\examples{
#For multichoice type questions

make_moostr(type="mchoice",
            ans=c("A two-sample t-test",
                  "A paired t-test",
                  "A one-sample t-test"),
            reward=c(50, 100, 0))

#For numerical type questions

make_moostr(type="num",
            ans=c(2, 2.1, 2.01),
            reward=c(30, 100, 50),
            tol = c(0, 0.1, 0.01))

make_moostr(type="num",
            ans=c(2, 2.1, 2.01),
            reward=c(50, 100, 50),
            tol = 0)

make_moostr(type="num",
            ans=c(2, 2.1, 2.01),
            reward=c(30, 100, 50))
}
